changed the router to App.js as well as style changes. Also added the AppBanner
